智能问答机器人-基于RAG技术/requirements.txt
langchain
langchain-community
dashscope
pypdf
faiss-cpu
python-dotenv
tiktoken